Let bundler manage LOAD_PATH
Bundler should add the correct LOAD_PATH based on each Gemfile.
When we manipulate the LOAD_PATH like that we end adding dependencies to it
that are not declared. This happens because Bundler will parse and execute all
the gemspecs in a source (`source ".." do`), even when it ends up ignorign it.