Add compatibility for attr_encrypted 4+ (#262)
This method assumes that the class is using attr_encrypted for its
encryption. That gem introduced a breaking change in its most recent
major version upgrade to help with Rails 7.0 compatibility. This ensures
both versions of attr_encrypted will be compatible with data_taster and
clarifies the intention for consumers that want to use this method.

This may be compatible with ActiveRecord encryption but is untested.