[FVR-299] Start inheriting Power Renovate config (#304)
Part of [FVR-299](https://runway.powerhrg.com/backlog_items/FVR-299)

This will ensure Power-standard Renovate config options like tagging
Renovate PRs with the 'dependencies' label, ignoring updates to the
`.github/workflows/stale.yml` file, and other important Renovate config
options are respected in this repository.

This includes removing the config entry to tag PRs with 'dependencies',
as this will be inherited from the parent config, as well as the
duplicate timezone definition. It also modifies the formatting of the
`matchUpdateTypes` array to match that of `extends` for consistency,
without any content changes.