[HFH-2834] Update serialize to pass keyword args (#308)
#### DEPRECATION WARNING: 
Passing the class as positional argument is deprecated and will be
removed in Rails 7.2.
Please pass the class as a keyword argument:
`serialize :options, type: Hash/Array/` etc


"DEPRECATION WARNING: Passing the class as positional argument is
deprecated and will be removed in Rails 7.2" in Rails 7.1, you need to
update the way you use the serialize method. Instead of passing the
class as a positional argument, you should pass it as a keyword argument


[Stackoverflow](https://stackoverflow.com/questions/79287986/warning-passing-the-class-as-positional-argument-is-deprecated-and-will-be-rem)
Introduced in 7.1.5.1
[Source](https://github.com/rails/rails/blob/v7.1.5.1/activerecord/lib/active_record/attribute_methods/serialization.rb#L183)

---------

Co-authored-by: Carlos Palhares <chjunior@gmail.com>