Bearer integration (#269)
This is a POC for [bearer](https://docs.bearer.com/), which is a static
scanner for security/privacy vulnerabilities.

While [brakeman](https://brakemanscanner.org/) is Rails-specific, Bearer
adds support for vulnerabilities on the language-level and supports the
following languages:

	1. JavaScript/TypeScript
	2. Ruby
	3. PHP
	4. Java
	5. Go
	6. Python

This PR aims to prove out bearer's effectiveness in the power-tools
repo.

Example failure:

![Screenshot 2024-11-06 at 9 56
43 AM](https://github.com/user-attachments/assets/6d99c673-26e9-4ce5-b03b-44ea3a5e3525)