Replace scoped method to work with Ruby 3.3 and Rails 6.1 (#356)
During Tempo Ruby 3.2 upgrade we started having the following error when
calling the `to` method:

```shell
ArgumentError:
       wrong number of arguments (given 1, expected 0; required keyword: subject)
``` 


This is because Rails 6.1 doesn't play well with the new way to pass
positional and keyword arguments to methods in Ruby 3 for scoped
methods.

In order to make that work for with all Rails versions, we need to
define the `to` method to be a class method instead of scoped one.

This won't affect other places that are using this method because the
signature of the method isn't changed.

---------

Co-authored-by: Carlos Palhares <carlos.palhares@powerhrg.com>